ultima noapte de dragoste

+++ Textbox.h
@@ -69,7 +69,7 @@ public:
                     if(text.str().length()<=limit){
                         inputLogic(charTyped);
                     }
-                    else if(charTyped==DELETE_KEY)
+                    else if(text.str().length()>limit&&charTyped==DELETE_KEY)
                         deleteLastChar();
                 }
                 else
